chore: move API base URL to .env for secrity and environment config
