+++ lib/api/ArticleService.js
@@ -1,7 +1,10 @@
-import axios from 'axios';
+//import instance from './apiService';
 
+import axios from 'axios';
+// 게시글 관련 자체 개발 API 백엔드와 코드잇 제공 API 백엔드가 달라 임시로 자체 API 백엔드로 연결함
+// 추후, sprint10에서 통합 예정
 const instance = axios.create({
-  baseURL: 'https://buffso-pandamarket-api.onrender.com',
+  baseURL: process.env.ARTICLE_API_BASE_URL,
 });
 
 instance.interceptors.response.use(
@@ -18,7 +21,8 @@ instance.interceptors.response.use(
       console.error('API 요청 중 에러 발생:', error.message);
       errorMessage = error.message || '요청 중 문제가 발생했습니다. 다시 시도해주세요.';
     }
-    return Promise.reject({ message: errorMessage });
+    error.message = errorMessage;
+    return Promise.reject(error);
   }
 );
 
@@ -28,7 +32,7 @@ export const getArticleList = async ({ page, pageSize, orderBy = 'recent' }) => 
     const res = await instance.get('/articles', {
       params: { page, pageSize, orderBy },
     });
-   // console.log('Response data:', res);
+    //console.log('Response data:', res);
     return Array.isArray(res.list) ? res.list : [];
   } catch (error) {
     console.error('게시글 목록을 가져오는데 실패했습니다:', error);
